move input to bt pkg
