+++ bt/Input.go
@@ -1,4 +1,4 @@
-package input
+package bt
 
 import (
 	"encoding/binary"
